# GH-44052: [C++][Compute] Reduce the complexity of row segmenter