… cpp/src/arrow/acero/hash_aggregate_test.cc
… cpp/src/arrow/compute/row/grouper.h