@pitrou 2024-09-11
I'm curious, is the `SimpleKeySegmenter` useful for performance?